changed consts on Players
