+++ packages/suite-base/src/players/IterablePlayer/IterablePlayer.test.ts
@@ -10,12 +10,8 @@ import * as _ from "lodash-es";
 
 import { signal } from "@lichtblick/den/async";
 import { fromSec } from "@lichtblick/rostime";
-import {
-  MessageEvent,
-  PlayerCapabilities,
-  PlayerPresence,
-  PlayerState,
-} from "@lichtblick/suite-base/players/types";
+import { PlayerCapabilities } from "@lichtblick/suite-base/players/constants";
+import { MessageEvent, PlayerPresence, PlayerState } from "@lichtblick/suite-base/players/types";
 import { mockTopicSelection } from "@lichtblick/suite-base/test/mocks/mockTopicSelection";
 
 import {
