+++ packages/suite-base/src/players/RosbridgePlayer.ts
@@ -27,11 +27,11 @@ import { MessageReader as ROS2MessageReader } from "@lichtblick/rosmsg2-serializ
 import { Time, fromMillis, toSec } from "@lichtblick/rostime";
 import { ParameterValue } from "@lichtblick/suite";
 import PlayerProblemManager from "@lichtblick/suite-base/players/PlayerProblemManager";
+import { PlayerCapabilities } from "@lichtblick/suite-base/players/constants";
 import {
   AdvertiseOptions,
   MessageEvent,
   Player,
-  PlayerCapabilities,
   PlayerState,
   PublishPayload,
   SubscribePayload,
