+++ packages/suite-base/src/testing/builders/PlayerBuilder.ts
@@ -1,10 +1,10 @@
 // SPDX-FileCopyrightText: Copyright (C) 2023-2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
 // SPDX-License-Identifier: MPL-2.0
 
+import { PlayerCapabilities } from "@lichtblick/suite-base/players/constants";
 import {
   BlockCache,
   MessageBlock,
-  PlayerCapabilities,
   PlayerPresence,
   PlayerState,
   PlayerStateActiveData,
